+++ src/python.rs
@@ -428,7 +428,7 @@ impl<'p> Python<'p> {
         unsafe {
             let mptr = ffi::PyImport_AddModule("__main__\0".as_ptr() as *const _);
             if mptr.is_null() {
-                return Err(PyErr::api_call_failed(self));
+                return Err(PyErr::fetch_last_error(self));
             }
 
             let globals = globals
@@ -438,7 +438,7 @@ impl<'p> Python<'p> {
 
             let code_obj = ffi::Py_CompileString(code.as_ptr(), "<string>\0".as_ptr() as _, start);
             if code_obj.is_null() {
-                return Err(PyErr::api_call_failed(self));
+                return Err(PyErr::fetch_last_error(self));
             }
             let res_ptr = ffi::PyEval_EvalCode(code_obj, globals, locals);
             ffi::Py_DECREF(code_obj);